feat: smart button see actions in action plan
