+++ onsp_co2/models/__init__.py
@@ -1,4 +1,5 @@
 # from . import carbon_general_mixin
+from . import common_mixin
 from . import carbon_copy_mixin
 from . import carbon_line_origin
 from . import carbon_mixin
