+++ onsp_co2/models/sustainability_scenario.py
@@ -4,7 +4,7 @@ from odoo import _, fields, models
 class SustainabilityScenario(models.Model):
     _name = "sustainability.scenario"
     _description = "Sustainability Scenario"
-    _inherit = ["mail.thread", "mail.activity.mixin"]
+    _inherit = ["mail.thread", "mail.activity.mixin", "common.mixin"]
 
     name = fields.Char(required=True)
     description = fields.Text(required=True)
@@ -25,33 +25,6 @@ class SustainabilityScenario(models.Model):
                 [("action_plan_ids.scenario_id", "=", scenario.id)]
             )
             scenario.action_qty = len(actions)
-
-    def _generate_action(self, model: str, title: str, ids: list[int]) -> dict:
-        """
-        Generate an action dictionary for the specified model and title.
-
-        This function creates an action dictionary that can be used to open a new window in the Odoo UI. The new window will display the specified model's data, filtered by the carbon domain.
-
-        Args:
-            model (str): The name of the model to display in the new window.
-            title (str): The title to display in the new window.
-            ids (list[int]): A list of integer IDs representing the records to be displayed in the new window.
-
-        Returns:
-            dict: An action dictionary that can be used to open a new window in the Odoo UI.
-        """
-        self.ensure_one()
-        return {
-            "name": _("%s %s", title, self.display_name),
-            "type": "ir.actions.act_window",
-            "res_model": model,
-            "views": [(False, "tree"), (False, "form")],
-            "domain": [("id", "in", ids)],
-            "target": "current",
-            "context": {
-                **self.env.context,
-            },
-        }
 
     def _get_action_plan_ids(self):
         return (
