+++ onsp_co2/views/sustainability_action_plan.xml
@@ -33,6 +33,31 @@
         <form>
           <field name="active" invisible="1" />
           <sheet>
+            <div class="oe_button_box" name="button_box">
+              <button
+                type="object"
+                name="action_see_actions"
+                class="oe_stat_button"
+                icon="fa-leaf"
+              >
+                <div class="o_field_widget o_stat_info">
+                  <span
+                    class="o_stat_value"
+                    attrs="{'invisible': [('action_qty', '&lt;=', 1)]}"
+                  >
+                    <field name="action_qty" widget="statinfo" nolabel="1" />
+                    Actions
+                  </span>
+                  <span
+                    class="o_stat_value"
+                    attrs="{'invisible': [('action_qty', '&gt;', 1)]}"
+                  >
+                    <field name="action_qty" widget="statinfo" nolabel="1" />
+                    Action
+                  </span>
+                </div>
+              </button>
+            </div>
             <widget
               name="web_ribbon"
               title="Archived"
